backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
qrcode[pil]==7.4.2
Pillow==10.1.0
yt-dlp==2023.11.16
python-magic==0.4.27
aiofiles==23.2.1
pydantic==2.5.0
starlette==0.27.0
